final? version
rename subscription.py to core.py because it contains the main part of the module (the MediaSub class)
Add a status to sources, to detect if they are down or up (sources need to raise SourceDown in order to detect if a source is down)
Add a without_subscription mode for source, then an httpx connection will be setup automatically. (Otherwise it will do it also, but it will log a warning message.
Add support for history using sqlite database.
Add a normalized_name for content. The goal is to avoid duplication, eg "One Piece" and "one-piece" will both be named as "one_piece" and the new content is only send 1 time.
^ Add an helper function "normalize" to perform string transformation.